Inherit from SwipeTableViewController.
