+++ Todoey/Controllers/SwipeTableViewController.swift
@@ -1,0 +1,57 @@
+//
+//  SwipeTableViewController.swift
+//  Todoey
+//
+//  Created by Jeff Ma on 13/5/20.
+//  Copyright © 2020 Jeff Ma. All rights reserved.
+//
+
+import UIKit
+import SwipeCellKit
+
+class SwipeTableViewController: UITableViewController, SwipeTableViewCellDelegate {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        tableView.rowHeight = 80
+    }
+
+    // MARK: - Table view data source
+    
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath) as! SwipeTableViewCell
+        cell.delegate = self
+        
+        return cell
+    }
+    
+    //MARK - Swipe Cell Delegate Methods
+
+    func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> [SwipeAction]? {
+        guard orientation == .right else { return nil }
+
+        let deleteAction = SwipeAction(style: .destructive, title: "Delete") { action, indexPath in
+            // handle action by updating model with deletion
+            self.updateModel(at: indexPath)
+        }
+
+        // customize the action appearance
+        deleteAction.image = UIImage(named: "delete-icon")
+
+        return [deleteAction]
+    }
+    
+    func tableView(_ tableView: UITableView, editActionsOptionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> SwipeOptions {
+        var options = SwipeOptions()
+        options.expansionStyle = .destructive
+        return options
+    }
+    
+    func updateModel(at indexPath: IndexPath) {
+        
+    }
+
+}
+
+
